requirements.txt
langchain==0.1.0
langchain-community==0.0.10
openai==1.6.1
sentence-transformers==2.2.2
chromadb==0.4.22
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
requests==2.31.0
fastapi==0.108.0
uvicorn==0.25.0
python-multipart==0.0.6
streamlit==1.29.0
numpy==1.24.3
pandas==2.0.3